requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
easyocr==1.7.1
et-xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.17.0
fsspec==2025.2.0
gitdb==4.0.12
gitpython==3.1.44
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy-loader==0.4
levenshtein==0.23.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.26.0
networkx==3.2.1
ninja==1.11.1.3
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pluggy==1.5.0
protobuf==4.25.6
pyarrow==19.0.0
pyclipper==1.3.0.post6
pydeck==0.9.1
pygments==2.19.1
pytest==8.0.2
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-levenshtein==0.23.0
pytz==2025.1
pyyaml==6.0.2
rapidfuzz==3.12.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.24.0
scipy==1.13.1
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
smmap==5.0.2
streamlit==1.32.1
sympy==1.13.1
tenacity==8.5.0
tifffile==2024.8.30
toml==0.10.2
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
triton==3.2.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
